wip(#1104): extract lower/expr_call/object_static
